[Silver III] Title: 주유소, Time: 328 ms, Memory: 35876 KB, Score: 100 point -EuiseonHub

+++ algorithm/baekjoon/Silver/13305. 주유소/주유소.java
@@ -9,19 +9,19 @@ public class Main {
         int cityCount = Integer.parseInt(br.readLine()) - 1;
         StringTokenizer st1 = new StringTokenizer(br.readLine());
         StringTokenizer st2 = new StringTokenizer(br.readLine());
-        int[] arrDistance = new int[cityCount];
-        int[] arrOilPrice = new int[cityCount];
-        int totalDistance = 0;
+        long[] arrDistance = new long[cityCount];
+        long[] arrOilPrice = new long[cityCount];
+        long totalDistance = 0;
         for (int i = 0; i < cityCount; i++) {
-            int tempDistance = Integer.parseInt(st1.nextToken());
+            long tempDistance = Integer.parseInt(st1.nextToken());
             arrDistance[i] = tempDistance;
             totalDistance += tempDistance;
             arrOilPrice[i] = Integer.parseInt(st2.nextToken());
         }
         int idx = 0;
-        int boughtDistance = arrDistance[0];
-        int currentOilPrice = arrOilPrice[0];
-        int boughtPrice = arrDistance[0] * currentOilPrice;
+        long boughtDistance = arrDistance[0];
+        long currentOilPrice = arrOilPrice[0];
+        long boughtPrice = arrDistance[0] * currentOilPrice;
         while (boughtDistance != totalDistance) {
             if (currentOilPrice > arrOilPrice[++idx]) {
                 currentOilPrice = arrOilPrice[idx];
